easter egg, content, readme, db
The easter egg camera feature implementation is implemented and working. New content is added from archives and existing content updated to approved versions in some cases. Readme is updated with changed to the database. Database is updated, including matching SQL files for path errors and changes to the geolocation coordinates. - theresa 4/9/2025 sprint 4 11:07 am